Add 'or 0'
This will prevent an error if we manually delete all the line items
from an order, by making sure that this sets the order total to zero,
instead of none.
Without this, the next line would cause an error, because it would
try to determine if none is less than or equal to the delivery threshold.